refactor(daemon): move bounded HTTPS transport into service
